Move non-template function defs to cpp files
Fixes linker errors.
