+++ test.hpp
@@ -191,18 +191,6 @@ class Test {
 };
 
 // https://stackoverflow.com/questions/111928/is-there-a-printf-converter-to-print-in-binary-format
-void printBits(size_t const size, void const* const ptr) {
-  unsigned char* b = (unsigned char*)ptr;
-  unsigned char byte;
-  int i, j;
-
-  for (i = size - 1; i >= 0; i--) {
-    for (j = 7; j >= 0; j--) {
-      byte = (b[i] >> j) & 1;
-      printf("%d", byte);
-    }
-  }
-  puts("");
-}
+void printBits(size_t const size, void const* const ptr);
 
 #endif
