+++ uTensor_util.hpp
@@ -6,28 +6,8 @@
 
 // #define MAX(A, B) ((A > B)? A:B)
 
+void return_error(int ret_val);
 #if MBED_CONF_APP_DEBUG_MSG
-void return_error(int ret_val) {
-  if (ret_val) {
-    printf(" [**Failure**] %d\r\n", ret_val);
-    printf("Exiting...\r\n");
-    fflush(stdout);
-    exit(-1);
-  } else {
-    printf("  [DONE]\r\n");
-  }
-}
-
-// void errno_error(void* ret_val) {
-//   if (ret_val == NULL) {
-//     printf(" [**Failure**] %d \r\n", errno);
-//     printf("Exiting...\r\n");
-//     fflush(stdout);
-//     exit(-1);
-//   } else {
-//     printf("  [DONE]\r\n");
-//   }
-// }
 
 #define ON_ERR(FUNC, MSG) \
   {                       \
@@ -47,7 +27,7 @@ void return_error(int ret_val) {
 // void errno_error(void* ret_val) { /*DOES NOTHING*/
 // }
 
-#define ON_ERR(FUNC, MSG) FUNC
+#define ON_ERR(FUNC, MSG) { FUNC; }
 #define DEBUG(MSG, ...)
 
 #endif
@@ -62,48 +42,10 @@ void return_error(int ret_val) {
 
 typedef std::vector<uint32_t> Shape;
 
-void printVector(std::vector<uint32_t> vec) {
-  printf("vector: \r\n");
-  for (uint32_t i : vec) {
-    printf("%d ", (unsigned int)i);
-  }
+void printVector(std::vector<uint32_t> vec);
+uint32_t htonl(uint32_t& val);
+uint16_t ntoh16(uint16_t val);
+uint32_t ntoh32(uint32_t val);
 
-  printf("\r\n");
-}
-
-// little endian to big endian
-uint32_t htonl(uint32_t& val) {
-  const uint32_t mask = 0b11111111;
-  uint32_t ret = 0;
-
-  ret |= val >> 24;
-  ret |= (val & (mask << 16)) >> 8;
-  ret |= (val & (mask << 8)) << 8;
-  ret |= val << 24;
-
-  return ret;
-}
-
-// big endian to little endian
-uint16_t ntoh16(uint16_t val) {
-  uint16_t ret = 0;
-
-  ret |= val >> 8;
-  ret |= val << 8;
-
-  return ret;
-}
-
-uint32_t ntoh32(uint32_t val) {
-  const uint32_t mask = 0b11111111;
-  uint32_t ret = 0;
-
-  ret |= val >> 24;
-  ret |= (val & (mask << 16)) >> 8;
-  ret |= (val & (mask << 8)) << 8;
-  ret |= val << 24;
-
-  return ret;
-}
 
 #endif
